+++ assets/js/theme/utils/ProgressButton.js
@@ -34,4 +34,5 @@ export default class ProgressButton {
       .next('.icon-spinner')
       .remove();
   }
+
 }